requirements.txt
accelerate
transformers==4.27.1
torch
datasets
huggingface_hub
jinja2
sentencepiece
protobuf
scikit-learn
isort==5.8.0
pyyaml>=5
plotly
requests
pandas==1.5.3
py7zr
hydra-core==1.3.2
pytimedinput==2.0.1